Help! This code has become super icky!
Ok, so there's a lot of changes here, but first off, a disclaimer: this update really needs more thorough testing to make sure I haven't broken backwards compatibility with mutperiod or the Drosophila data. Unfortunately, I really don't want to do that right now. A smarter me would set up unit tests, but here we are. Anyway, the changes I made (I think this is all of them) are as follows:
- Added a new supplemental information handler for pulling out information on specific columns and adding it to the output.
- Added a new function, addCustomSupplementalInformationHandler, because the new handler (mentioned above) takes more inputs than its parent class, so it needs a different wrapper function that doesn't assume default constructor inputs in order to add it to the counter output data handler. Big yikes on the code architecture there. Someone should probably fix that.
- Standardized the default input structures to split specifically on tabs instead of using default splitting.
- Added a new parameter to the outputDataWriter that allows for masking of the final counts. This is really helpful in the event that you mostly just want to check for encompassment and maybe want to add some supplemental information or something.