Create RemoveBlacklistedRegions.py
This script takes a removes blacklisted regions from files of genomic features, like nucleosomes. I decided not to use the usual CountThisInThat paradigm because it actually treats encompassed features that span regions as the midpoint of that region. Thus, it wouldn't detect overlaps properly for features that span regions of different sizes (e.g., genes).